Added line drawing algorithm to build modified
Added a naive line drawing algorithm that takes two coordinates
and returns a list of all coords that lie on the line between the
two coordinates. To draw a line you have to hold down middle mouse
button on the start tile and let go on the tile you want the line to
end on. A preview of the line is shown.