Added speed controls and working playmode
I added a button to control the speed of the algorithm.
It has 3 stages and a fourth one that is unused atm. You can cycle
through the stages by left clicking on the speed button.
The speeds of each stage are dictated by the list speedtimes.
The program outputs 120 fps, so a speed of 40 would be equal to
120/40 = 3 updates per second.
The speed can be changed at any point in time, even while the algorithm
is running.
Need to add a check whether goal and start are placed before you can
click start.
Going back into build mode resets the algorithm.
I need to add a rest button once the algo is finished and a stop button
to stop algo while its running.