Added functional stop button and made new icons
I added a functional stop button, that can be used to stop the
algorithm while its running. It works identical to the replay button
which is a stage of the playbutton, but it only appears once the
algorithm is started and disappears, once its stopped again.
Pressing the stop button only resets the algorithm and the stopgobutton.
The speed setting are not affected.

I also made pixelart icons to replace the placeholder once. Only the
build mode button is missing. They look pretty sick, but they are not
completely uniform, which is triggering my ocd a bit. Padding to edges
is not the same for all of them, might change in the future.