Added helpbutton, improved popups, added text
I added a help button that si supposed to provide help.
So far it does nothing.
I also added a text renderer, which allows you to render text to the
screen. So far only single line.
I added a topbar to popups that you can click and drag to move the
popup. If you move the popup outside the window, it gets reset to the
middle.
I also restructured the popup code in the while loop.