Made popupmoving a class attribute + changed icons
There used to be a global boolean popupmoving that was true when
the popup was currently being moved, false otherwise.
This has now been turned into a Popup Class attribute
That would allow for multiple seperate popups if needed.

Also added a few alternative versions of the icons, will have to
decide later on.