Added multiple bugs (linedrawing and timer)
If a line was currently being drawn and the user clicked on any of the
buttons while holding down mouse 2, the line would bug out.
I made it so the line gets reset whenever the user clicks on any buttons
while drawing a line.
I also fixed a small bug with the timer in algo mode.
If the user changed the algo speed before clicking on play, the first
iteration would always run at the default speed and the speed would only
be adjusted after the first iteration. Now its changed immedeatly.