Added support for multiple different popup objects
So far there was only one single popup object called popup1.
Now, you can create as many as you want and they all render correctly.
This is useful, because popups are used in many different usecases and
modifying popup1 every time would be difficult.
There can only be one active popup at any point in time, but they all
work as intended.
Also added support for text, that is automaticall displayed in the
popups. You can choose the text and it auto formats to fit into
the popup. It can extend over the bottom of the popup if its too long.