Improved the runtime of dfs significantly
Before:
DFS has to check if the found tile has already been found before.
I used to do this by entering each tiles into a dictionary
where the tile was the key and the parent tile was the value.
To check if a tile has already been found, I tested if tile was in
dictionary.keys(). This operation cost O(n^2) I believe. (n being the
gridsize).
I also used to enter the entire list of dictionary.keys() and the entire
stack into the main matrix to display the tiles as green on each
iteration. This was incredibly stupid and was also the main problem.

Now:
I have now implemented a visited_matrix that is the same size as the
main matrix that tracks the grid, and contains boolean values.
If a node has been visited, matrix[node] = True, else False. Entering
and checking data this way should be O(1) as far as I know...
I also made it so that dfs step returns the newly visited node and the
newly added stack items. This way adding those to the main matrix should
only be O(1).

On a empty 64x64 matrix, times have improved from 137secs to 43secs
(Theoretical minimum should be around 34 seconds, but I dont think its
achievable, because rendering takes time. Could be possible though)