Added a new Popup class with ok button
I added a new class that is a popup with a simple ok button.

I also added a new popup for when the user tries to swicht into algo
mode, but hasnt place the start or goal tile yet.
This prevents the program from crashing in algo mode because there is no
start or goal.