Learned how inheritance works and cut out junk
I knew my previous implementation of the Popup classes was trash,
because I couldn't be bothered to look up how to inherit properly.
I simply copied the enitre function of the base class over and added
one or two new attributes. I replaced all of that with super() calls.
Its about as short as it can be now and seems to work fine.