Added BFS and the program now shows finished path
I added the option to choose between BFS and DFS. They appear to work
as inteded. I still need to cover some edgecases liek switching algos
mid run, that might cause the game to crash.
I also added a skip button that runs the algorun function in a while
loop until algo is finished. Useful, because even the fastest speed only
finds a 120 new tiles per second, which means that 64x64 grid would take
30 seconds to clear. I know how to change that, but right now I cant be
bothered.
Also added a function that draws the path to the goal if one was found.