Added functioning Dijkstra Algo
I added support for dijkstras algorithm. I believe it works like it
should, there might be some bugs. Since all edge wheights are 1,
it works just like bfs, but the priority queue orders the vertices
that have the same cost differently, so the found path can differ.
The length of the path will still be the same, its just a different
shortest path.